requirements.txt.txt
python-telegram-bot==20.7
python-dotenv
pandas
openpyxl